+++ src/ContaoManager/Plugin.php
@@ -35,7 +35,7 @@ class Plugin implements BundlePluginInterface, ConfigPluginInterface
      */
     public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
     {
-        $loader->load('@HeimrichHannotContaoEncoreBundle/Resources/config/config.yml');
-        $loader->load('@HeimrichHannotContaoEncoreBundle/Resources/config/services.yml');
+        $loader->load('@HeimrichHannotContaoEncoreBundle/config/config.yml');
+        $loader->load('@HeimrichHannotContaoEncoreBundle/config/services.yml');
     }
 }