+++ src/Resources/assets/js/encore.empty-entry.js
@@ -1,3 +1,0 @@
-/**
- this file has been included by heimrichhannot/contao-encore-bundle because the command line parameter entries had been set
- **/